backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
django-filter==23.5
Pillow==10.1.0
python-decouple==3.8
drf-spectacular==0.26.5
django-extensions==3.2.3